requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2

# Database
psycopg[binary]==3.2.1
asyncpg==0.29.0

# Vector search
pgvector==0.3.2

# AI/ML
openai==1.51.0
anthropic==0.34.2

# HTTP + XML parsing (eCFR API)
httpx==0.27.2
lxml==5.3.0

# Utilities
rich==13.8.1
pyyaml==6.0.2
numpy==1.26.4
